Add failing spec to demonstrate Property#unique_index :key bug.
At issue is that a Property which is a member of a Composite Primary Key which is also a member of a separate unique index will break the :key unique index. This is because such a double-unique index Property will only appear in the non-CPK unique index, which means that the remaining components of the CPK will be used to form yet another unique index (other than the CPK and the declared unique index).